+++ install.sh
@@ -49,25 +49,6 @@ if [ ! x"$SUDO_USER" = x"" ]; then
 	chown "$SUDO_USER:" hikka-install.log
 fi
 
-if [ ! x"" = x"$DYNO" ] && ! command -v python >/dev/null; then
-	# We are running in a heroku dyno without python, time to get ugly!
-	runout git clone https://github.com/heroku/heroku-buildpack-python || {
-		printf "Bootstrap download failed!"
-		exit 1
-	}
-	rm -rf .heroku .cache .profile.d requirements.txt runtime.txt .env
-	mkdir .cache .env
-	echo "python-3.9.6" >runtime.txt
-	echo "pip" >requirements.txt
-	STACK=heroku-18 runout bash heroku-buildpack-python/bin/compile /app /app/.cache /app/.env ||
-		{
-			printf "Bootstrap install failed!"
-			exit 1
-		}
-	rm -rf .cache
-	export PATH="/app/.heroku/python/bin:$PATH" # Prefer the bootstrapped python, incl. pip, over the system one.
-fi
-
 if [ -d "Hikka/hikka" ]; then
 	cd Hikka || {
 		printf "\rError: Install git package and re-run installer"
@@ -75,8 +56,8 @@ if [ -d "Hikka/hikka" ]; then
 	}
 	DIR_CHANGED="yes"
 fi
-if [ -f ".setup_complete" ] || [ -d "hikka" -a ! x"" = x"$DYNO" ]; then
-	# If hikka is already installed by this script, or its in Heroku and installed
+if [ -f ".setup_complete" ]; then
+	# If hikka is already installed by this script
 	PYVER=""
 	if echo "$OSTYPE" | grep -qE '^linux-gnu.*'; then
 		PYVER="3"